WEBOPS-2240 Changed project structure to introduce blueprint
